% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{adjust_asciidoc_hooks}
\alias{adjust_asciidoc_hooks}
\title{Adjust \pkg{knitr}'s Hooks for `asciidoc`}
\usage{
adjust_asciidoc_hooks(which = c("message", "error", "warning"),
  replacement = "source")
}
\arguments{
\item{which}{Character vector naming the output hooks to be replaced by
the \code{replacement}'s hooks.}

\item{replacement}{The hook with which to replace the hooks given by
\code{which}.}
}
\value{
The return value of
\code{\link[knitr:knit_hooks]{knitr::knit_hooks$set}},
\code{\link[base:invisible]{invisibly}} \code{\link{NULL}}, currently.
}
\description{
By default, \pkg{knitr} renders messages, warnings and errors to
[NOTE|WARNING|ERROR]-blocks in `asciidoc`, which is ... not my choice.
}
\details{
This is a modified version of
\code{\link[knitr:render_asciidoc]{knitr::render_asciidoc}} of  \pkg{knitr}
version 1.18.7.
}
