% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciidoc.R
\name{asciidoc}
\alias{asciidoc}
\title{Run `asciidoc` on a File}
\usage{
asciidoc(file_name, ..., knit = NA)
}
\arguments{
\item{file_name}{The file to run `asciidoc` on.}

\item{...}{arguments passed to `.asciidoc`.}

\item{knit}{Knit the file first using \code{\link[knitr:knit]{knitr::knit}}?
If set to NA, the file's name is checked for the sequence `.R` and if found,
knit is set to TRUE. Set to anything apart from \code{NA} to avoid
auto-checking the file's name.}
}
\value{
The output's file names.
}
\description{
Wrapper to .asciidoc using html and slidy as output.
}
