% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_version_sufficient.R
\name{is_version_sufficient}
\alias{is_version_sufficient}
\title{Is a Version Requirement Met?}
\usage{
is_version_sufficient(installed, required)
}
\arguments{
\item{required}{The version required.}

\item{available}{The version available.}
}
\value{
\code{\link{TRUE}}, if so,  \code{\link{FALSE}} otherwise.
}
\description{
Just a wrapper to \code{\link{compareVersion}}, I regularly forget how to use
it.
}
\examples{
is_version_sufficient(installed = "1.0.0", required = "2.0.0")
is_version_sufficient(installed = "1.0.0", required = "1.0.0")
}
\keyword{internal}
