% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasciidoc-package.R
\docType{package}
\name{rasciidoc-package}
\alias{rasciidoc-package}
\title{Create Reports Using R and `asciidoc`}
\description{
Inspired by Karl Broman's reader on using knitr with asciidoc
(<http://kbroman.org/knitr_knutshell/pages/asciidoc.html>), this is
a wrapper to and a slight modification of knitr.
}
\details{
You will find the details in\cr
\code{vignette("An_Introduction_to_rasciidoc", package = "rasciidoc")}.
}
\keyword{package}
