% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{rasciidoc}
\alias{rasciidoc}
\title{Render an \command{asciidoc} File}
\usage{
rasciidoc(file_name, ...)
}
\arguments{
\item{file_name}{The file to run \command{asciidoc} on.}

\item{...}{arguments passed to \command{asciidoc} via \code{\link{system2}}.}
}
\value{
\code{\link[base:invisible]{Invisibly}} \command{asciidoc} return
value.
}
\description{
This is the basic interface to \command{asciidoc}. Not more than a call to
\code{\link{system2}} and checks on \command{asciidoc} and
\command{source-highlight}.
You should usually not call it directly, see
\code{\link{render}} and \code{\link{render_slides}} for wrappers.
}
\examples{
wd <- file.path(tempdir(), "rasciidoc")
dir.create(wd)
file  <- system.file("files", "minimal", "knit.asciidoc",
                     package = "rasciidoc")
file.copy(file, wd)
rasciidoc::rasciidoc(file.path(wd, basename(file)), "-b docbook")
dir(wd, full.names = TRUE)
unlink(wd, recursive = TRUE)
}
