% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{Spin or Knit and Render a `Rasciidoc` File}
\usage{
render(file_name, knit = NA, adjust_hooks = TRUE, envir = parent.frame(),
  ...)
}
\arguments{
\item{file_name}{The file to run `asciidoc` on.}

\item{knit}{Knit the file first using \code{\link[knitr:knit]{knitr::knit}}?
If set to \code{\link{NA}}, knitting is based on the file's contents or name.
Set to \code{\link{TRUE}}
to force knitting or to \code{\link{FALSE}} (anything apart from
\code{\link{TRUE}} or \code{\link{NA}}, really), to
disable knitting.}

\item{adjust_hooks}{Adjust knitr's output hooks for `asciidoc` files using
the defaults of \code{\link{adjust_asciidoc_hooks}}?}

\item{envir}{The frame in which to render.}

\item{...}{arguments passed to `asciidoc` via \code{\link{system2}}.}
}
\value{
The return value of \code{\link{rasciidoc}}.
}
\description{
Spin or Knit (if required) and render an `Rasciidoc` file.
}
