% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_slides}
\alias{render_slides}
\title{Knit and Render an `asciidoc` File to html and slidy}
\usage{
render_slides(file_name, knit = NA, adjust_hooks = TRUE)
}
\arguments{
\item{file_name}{The file to run `asciidoc` on.}

\item{knit}{Knit the file first using \code{\link[knitr:knit]{knitr::knit}}?
If set to \code{\link{NA}}, knitting is based on the file's contents or name.
Set to \code{\link{TRUE}}
to force knitting or to \code{\link{FALSE}} (anything apart from
\code{\link{TRUE}} or \code{\link{NA}}, really), to
disable knitting.}

\item{adjust_hooks}{Adjust knitr's output hooks for `asciidoc` files using
the defaults of \code{\link{adjust_asciidoc_hooks}}?}
}
\value{
The output's file names.
}
\description{
Knit and Render an `asciidoc` File to html and slidy
}
