% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciidoc-package.R
\docType{package}
\name{asciidoc-package}
\alias{asciidoc-package}
\title{Create Reports Using `knitr` and `asciidoc`}
\description{
I want to use asciidoc for html and slidy.
}
\details{
You will find the details in\cr
\code{vignette("An_Introduction_to_asciidoc", package = "asciidoc")}.
}
\keyword{package}
