% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{asciidoc}
\alias{asciidoc}
\title{Render an `asciidoc` File}
\usage{
asciidoc(file_name, ...)
}
\arguments{
\item{file_name}{The file to run `asciidoc` on.}

\item{...}{arguments passed to `asciidoc` via \code{\link{system2}}.}
}
\value{
\code{\link[base:invisible]{Invisibly}}`asciidoc`'s return value.
}
\description{
This is the basic interface to `asciidoc`. Not more than a call to
\code{\link{system2}} and checks on `asciidoc` and `source-highlight`.
}
