% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{Spin or Knit and Render a `Rasciidoc` File}
\usage{
render(file_name, knit = NA, envir = parent.frame(),
  working_directory = dirname(file_name), hooks = c("message", "error",
  "warning"), replacement = "source", ...)
}
\arguments{
\item{file_name}{The file to run `asciidoc` on.}

\item{knit}{Knit the file first using \code{\link[knitr:knit]{knitr::knit}}?
If set to \code{\link{NA}}, knitting is based on the file's contents or name.
Set to \code{\link{TRUE}}
to force knitting or to \code{\link{FALSE}} (anything apart from
\code{\link{TRUE}} or \code{\link{NA}}, really), to
disable knitting.}

\item{envir}{The frame in which to render.}

\item{working_directory}{Where to run \code{\link[knitr:knit]{knitr::knit}}
or \code{\link[knitr:spin]{knitr::spin}}, defaults to the input file's
directory to ensure that sourcing code into the input file works correctly.}

\item{hooks}{Character vector naming the output hooks to be replaced by
the \code{replacement}'s hooks.}

\item{replacement}{The hook with which to replace the hooks given by
\code{hooks}.}

\item{...}{arguments passed to `asciidoc` via \code{\link{system2}}.}
}
\value{
The return value of \code{\link{rasciidoc}}.
}
\description{
Spin or Knit (if required) and render an `Rasciidoc` file.
}
