% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{Knit and Render an `asciidoc` File}
\usage{
render(file_name, knit = NA, ...)
}
\arguments{
\item{file_name}{The file to run `asciidoc` on.}

\item{knit}{Knit the file first using \code{\link[knitr:knit]{knitr::knit}}?
If set to \code{\link{NA}}, knitting is based on the file's contents or name.
Set to \code{\link{TRUE}}
to force knitting or to \code{\link{FALSE}} (anything apart from 
\code{\link{TRUE}} or \code{\link{NA}}, really), to
disable knitting.}

\item{...}{arguments passed to `asciidoc` via \code{\link{system2}}.}
}
\value{
The return value of \code{\link{rasciidoc}}.
}
\description{
Knit (if required) and render an `asciidoc` file.
}
